requirements.txt
streamlit
python-dotenv
langchain-openai
langchain